ada: Fix computation of new size when reallocating unbounded string
The procedure Realloc_For_Chunk which is used to reallocate an unbounded
string when needed may lead in theory to an overflow, due to the use of
variable S_Length denoting the current allocated length instead of
Source.Last denoting the current string length. Now fixed.

This has no effect in practice since the only targets that use this
version of Ada.Strings.Unbounded do not have enough memory to make it
possible to have an overflow here.

gcc/ada/

	* libgnat/a-strunb.adb (Realloc_For_Chunk): Fix computation of new
	size.