ada: Fix detection of suspicious loop patterns
This patch fixes an assertion failure in some cases in the code to
warn about possible misuse of range attributes in loop. The root of
the problem is that this code failed to consider the case where the
outer loop is a while loop.

Also fix a typo in a nearby comment.

gcc/ada/

	* sem_ch5.adb (Analyze_Loop_Statement): Fix loop pattern detection
	code. Fix typo.