rs6000: Consider explicit VSX when masking off ALTIVEC [PR115688]
PR115688 exposes an inconsistent state in which we have VSX
enabled but ALTIVEC disabled.  There is one hunk:

  if (main_target_opt && !main_target_opt->x_rs6000_altivec_abi)
    rs6000_isa_flags &= ~((OPTION_MASK_VSX | OPTION_MASK_ALTIVEC)
			  & ~rs6000_isa_flags_explicit);

which disables both VSX and ALTIVEC together only considering
them explicitly set or not.  For the given case, VSX is explicitly
specified, altivec is implicitly enabled as it's part of set
ISA_2_6_MASKS_SERVER.  When falling into the above hunk, vsx is
kept as it's explicitly enabled but altivec gets masked off, it's
unexpected.

This patch is to consider explicit VSX when masking off ALTIVEC,
not mask off it if TARGET_VSX and it's explicitly set.

	PR target/115688

gcc/ChangeLog:

	* config/rs6000/rs6000.cc (rs6000_option_override_internal): Consider
	explicit VSX when masking off ALTIVEC.

gcc/testsuite/ChangeLog:

	* gcc.target/powerpc/pr115688.c: New test.