i386: Promote {QI,HI}mode x86_mov<mode>cc_0_m1_neg to SImode
Promote HImode x86_mov<mode>cc_0_m1_neg insn to SImode to avoid
redundant prefixes. Also promote QImode insn when TARGET_PROMOTE_QImode
is set. This is similar to promotable_binary_operator splitter, where we
promote the result to SImode.

Also correct insn condition for splitters to SImode of NEG and NOT
instructions. The sizes of QImode and SImode instructions are always
the same, so there is no need for optimize_insn_for_size bypass.

	gcc/ChangeLog:

	* config/i386/i386.md (x86_mov<mode>cc_0_m1_neg splitter to SImode):
	New splitter.
	(NEG and NOT splitter to SImode): Remove optimize_insn_for_size_p
	predicate from insn condition.