arm: clean up some legacy FPA related cruft.
Support for the FPA on Arm was removed after gcc-4.7, but this little
bit of crufty code was left behind.  In particular the code to support
the 'N' modifier in assembly code was left behind and this lead to a
trail of other code that depended on it, even though most of the
constants that it supported had been removed in the original cleanup.

This patch removes most of the remaining cruft and simplifies the one
bit that remains: to determine whether an RTL construct contains 0.0 we
don't need to convert it to a real value, we can simply compare it to
CONST0_RTX of the appropriate mode.

gcc/

	* config/arm/arm.cc (fp_consts_initited): Delete variable.
	(value_fp0): Likewise.
	(init_fp_table): Delete function.
	(fp_const_from_val): Likewise.
	(arm_const_double_rtx): Rework to avoid converting to REAL_VALUE_TYPE.
	(arm_print_operand, case 'N'): Make use of this case an error.