ada: Set DECL_NAMELESS in create_type_decl
When using minimal encodings, most artificial types do not need to
have their names emitted in the DWARF.  This patch changes the
compiler to generally omit these names.

However, a subset of names are needed: when the compiler creates an
artificial type for certain kinds of arrays, the name is needed by
gdb.  So, a new parameter is added to create_type_decl to allow this
omission to be disabled.

Note that simply passing 'false' as the artificial_p argument to
create_type_decl doesn't work properly -- other parts of the compiler
seem to rely on this flag being set, and so making this change causes
ICEs.

gcc/ada/ChangeLog:

	* gcc-interface/decl.cc (gnat_to_gnu_entity): Update some calls to
	create_type_decl.
	* gcc-interface/gigi.h (create_type_decl): Add can_be_nameless parameter.
	* gcc-interface/utils.cc (create_type_decl): Add can_be_nameless
	parameter.  Set DECL_NAMELESS on type decl.