rs6000, Remove __builtin_vec_set_v1ti, __builtin_vec_set_v2df, __builtin_vec_set_v2di
This patch removes the __builtin_vec_set_v1ti, __builtin_vec_set_v2df
and __builtin_vec_set_v2di built-ins.  The users should just use
normal C-code to update the various vector elements.  This change was
originally intended to be part of the earlier series of cleanup
patches.  It was initially thought that some additional work would be
needed to do some gimple generation instead of these built-ins.
However, the existing default code generation does produce the needed
code.    For the vec_set bif, the equivalent C code is as good or
better than the built-in.  For the vec_insert bif whose resolving
previously made use of the vec_set bif, the assembly code generation
is as good as before with the -O3 optimization.

Remove the built-ins, use the default gimple generation instead.

gcc/ChangeLog:
	* config/rs6000/rs6000-builtins.def (__builtin_vec_set_v1ti,
	__builtin_vec_set_v2df, __builtin_vec_set_v2di): Remove built-in
	definitions.
	* config/rs6000/rs6000-c.cc (resolve_vec_insert): Remove the
	handling for constant vec_insert position with
	VECTOR_UNIT_VSX_P V1TImode, V2DFmode and V2DImode modes.